Enable strict tsconfig option and fix errors

+++ todo-api/src/users/users.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UsersRepository } from './users.repository';
 import { User } from './entities/user.entity';
-import { strict as assert } from 'assert';
+import { ClassValidatorUtil } from 'src/utils/class-validator.util';
+import { ExceptionConstants } from 'src/constants/exception.constants';
 
 @Injectable()
 export class UsersService {
@@ -15,11 +16,11 @@ export class UsersService {
      * Creates a new user.
      * @param createUserDto - The data for creating a new user.
      * @returns A promise that resolves to the created user.
-     * @throws {AssertionError} If the CreateUserDto is not provided.
+     * @throws {ValidationException} If user email or password is invalid.
      * @throws {InternalServerErrorException} If the user creation fails.
      */
     async create(createUserDto: CreateUserDto): Promise<User> {
-        assert(createUserDto, 'CreateUserDto must be provided');
+        await ClassValidatorUtil.validate(createUserDto);
         return await this.usersRepository.insert(createUserDto);
     }
 
@@ -27,11 +28,14 @@ export class UsersService {
      * Finds a user by their ID.
      * @param id - The ID of the user to find.
      * @returns A promise that resolves to the found user.
-     * @throws {AssertionError} If the user ID is not provided.
+     * @throws {BadRequestException} If the user ID is less than 1.
      * @throws {NotFoundException} If the user is not found.
      */
     async findOneById(id: number): Promise<User> {
-        assert(id > 0, 'User ID must be greater than 0');
+        if (!id || id < 1) {
+            throw new BadRequestException(ExceptionConstants.INVALID_USER_ID);
+        }
+
         return await this.usersRepository.getOneById(id);
     }
 
@@ -39,10 +43,13 @@ export class UsersService {
      * Finds a user by their email.
      * @param email - The email of the user to find.
      * @returns A promise that resolves to the found user, or null if not found.
-     * @throws {AssertionError} If the email is not provided.
+     * @throws {BadRequestException} If the email is not provided.
      */
     async findOneByEmail(email: string): Promise<User | null> {
-        assert(email, 'Email must be provided');
+        if (!email) {
+            throw new BadRequestException(ExceptionConstants.INVALID_EMAIL);
+        }
+
         return await this.usersRepository.getOneByEmail(email);
     }
 
@@ -51,15 +58,17 @@ export class UsersService {
      * @param id - The ID of the user to update.
      * @param updateUserDto - The data for updating the user.
      * @returns A promise that resolves to the update result.
-     * @throws {AssertionError} If the user ID is not provided, or if the password or refresh token is not provided.
+     * @throws {BadRequestException} If the user ID is less than 1, or if no fields are provided to update.
      * @throws {NotFoundException} If the user to update is not found.
      */
     async update(id: number, updateUserDto: UpdateUserDto): Promise<User> {
-        assert(id > 0, 'User ID must be greater than 0');
-        assert(
-            updateUserDto.password || updateUserDto.refreshToken,
-            'At least one field must be provided to update user',
-        );
+        if (!id || id < 1) {
+            throw new BadRequestException(ExceptionConstants.INVALID_USER_ID);
+        }
+
+        if (!updateUserDto.password && !updateUserDto.refreshToken) {
+            throw new BadRequestException('At least one field must be provided to update user');
+        }
 
         return await this.usersRepository.update(id, updateUserDto);
     }
@@ -68,10 +77,13 @@ export class UsersService {
      * Nullifies the refresh token of a user.
      * @param id - The ID of the user.
      * @returns A Promise that resolves to the updated User object.
-     * @throws {AssertionError} Iff the user ID is not greater than 0.
+     * @throws {BadRequestException} If the user ID is less than 1.
      */
     async nullifyUserRefreshToken(id: number): Promise<User> {
-        assert(id > 0, 'User ID must be greater than 0');
+        if (!id || id < 1) {
+            throw new BadRequestException(ExceptionConstants.INVALID_USER_ID);
+        }
+
         return await this.usersRepository.nullifyUserRefreshToken(id);
     }
 
@@ -79,11 +91,14 @@ export class UsersService {
      * Removes a user.
      * @param id - The ID of the user to remove.
      * @returns A promise that resolves to the number of deleted rows.
-     * @throws {AssertionError} If the user ID is not provided.
+     * @throws {BadRequestException} If the user ID is less than 1.
      * @throws {NotFoundException} If the user to delete is not found.
      */
     async remove(id: number) {
-        assert(id > 0, 'User ID must be greater than 0');
+        if (!id || id < 1) {
+            throw new BadRequestException(ExceptionConstants.INVALID_USER_ID);
+        }
+
         return await this.usersRepository.delete(id);
     }
 }
